2026-07-31 mc-161607
- Engine

  Add support for web (via Emscripten)
  Add VirtualFileSystem and NativeFileSystem class.
  Add support for WebFileSystem.
  Fix finish bug in Application.
  Change mouse input code. Register  mouse movement instead of the mouse
  position.
  Add lgfx includes replaced by 'Graphics/GL/GL.h'
  All classes that open files go through the VirtualFileSystem.

- Game

  Mount filesystems on init.
  All classes that open files go through the VirtualFileSystem.
  Change Rubydung.cpp to Minecraft.cpp.
  Add download assets at init (for web systems)

+++ game/Renderer/ZombieRenderer.cpp
@@ -25,7 +25,7 @@ void ZombieRenderer::Render(Level &lev, Player &player, float alpha, float secon
 	m_Shader->SetFloat("s_fdensity0", GlobalGP.fg0.density);
 	m_Shader->SetVec4("s_fcolor1", GlobalGP.fg1.color);
 	m_Shader->SetFloat("s_fdensity1", GlobalGP.fg1.density);
-	m_Shader->SetFloat("s_seconds", fmod(seconds, 2 * M_PI));
+	m_Shader->SetFloat("s_seconds", (float)fmod(seconds, 2 * M_PI));
 	m_Shader->SetFloat("s_scale", 0.058333334f);
 
 	std::vector<ZombieInstance>().swap(m_Instances);
